solitare: weaken "discarding must be followed by discarding the same suit or using the stack" -- must be making an empty stack
Was fairly easy to get into situations where you had to discard another card to discard a card of the same suit.
Consider the end of the game, you might have to discard an unrelated card in a different stack in order to discard the next card of the same suit.